+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 
 ## [Unreleased]
 
+### Breaking
+
+- Marpit requires Node >= 8.
+
 ### Added
 
 - Add [documentation of fragmented list](https://marpit.marp.app/fragmented-list) ([#152](https://github.com/marp-team/marpit/pull/152))
+
+### Removed
+
+- Drop support for Node 6.x ([#139](https://github.com/marp-team/marpit/issues/139), [#155](https://github.com/marp-team/marpit/pull/155))
+- Remove deprecated constructor options: `backgroundSyntax`, `filters`, `inlineStyle`, and `scopedStyle` ([#156](https://github.com/marp-team/marpit/pull/156))
 
 ## v0.9.2 - 2019-04-08